requirements.txt
python-telegram-bot==20.6
Flask
gspread
oauth2client
openai
nest_asyncio